refactor web Game component
