+++ packages/web/src/utils/isActionPhaseActive.ts
@@ -1,0 +1,7 @@
+export function isActionPhaseActive(
+  currentPhase: string,
+  actionPhaseId: string | undefined,
+  tabsEnabled: boolean,
+): boolean {
+  return tabsEnabled && currentPhase === actionPhaseId;
+}
